+++ shared/src/lib.rs
@@ -1,5 +1,5 @@
 #[derive(Debug, Clone, serde::Deserialize, serde::Serialize)]
 pub enum EguiCommand {
     Ping,
-    HighlightMember(i32),
+    HighlightMember(i64),
 }
